Feat : add modal
